1-string_nconcat.c

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -11,7 +11,7 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 	char *ptr;
-	int len1, len2, i, j;
+	size_t len1, len2, i, j;
 
 	for (len1 = 0; s1[len1]; len1++)
 		;
